+++ src/style/style.ts
@@ -1908,7 +1908,7 @@ export class Style extends Evented<MapEventType> {
         let placementCommitted = false;
 
         const layerTiles = {};
-        const reindexedBuckets = new Set<number>();
+        const reindexedBucketIds = new Set<number>();
 
         for (const layerID of this._order) {
             const styleLayer = this._layers[layerID];
@@ -1923,8 +1923,8 @@ export class Style extends Evented<MapEventType> {
 
             const layerResult = this.crossTileSymbolIndex.addLayer(styleLayer, layerTiles[styleLayer.source], transform.center.lng);
             symbolBucketsChanged ||= layerResult.bucketsChanged;
-            for (const bucketInstanceId of layerResult.reindexedBuckets) {
-                reindexedBuckets.add(bucketInstanceId);
+            for (const bucketInstanceId of layerResult.reindexedBucketIds) {
+                reindexedBucketIds.add(bucketInstanceId);
             }
         }
         this.crossTileSymbolIndex.pruneUnusedLayers(this._order);
@@ -1975,7 +1975,7 @@ export class Style extends Evented<MapEventType> {
             for (const layerID of this._order) {
                 const styleLayer = this._layers[layerID];
                 if (styleLayer.type !== 'symbol') continue;
-                this.placement.updateLayerOpacities(styleLayer, layerTiles[styleLayer.source], placementCommitted ? null : reindexedBuckets);
+                this.placement.updateLayerOpacities(styleLayer, layerTiles[styleLayer.source], reindexedBucketIds);
             }
         }